backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
slowapi>=0.1.9
sse-starlette>=2.1.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Development/Linting dependencies
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
